Add listToken attribute to items in database
Overview: this commit makes sure that users can add items to a specific list and they’re only shown items from the list they are currently managing.

Details: Previously there were no attributes being create on items that would signify which list those items belong to, so users were shown all items in the database — yikes!

In this commit, we’ve added a listToken attribute to each item that gets stored in the database. The listToken is pulled from localStorage when the addItem function runs.

Then, in the List component, we’ve added a filter to our Firestore query that will only display items from the database if that item’s listToken matches the token that is currently stored in localStorage.