Add Cypress to project for end-to-end testing
I've added Cypress to the project for end-to-end testing because after reading through the intro of their documentation and implementing some tests, I think this is a really great tool for folks who are new to testing.

It allows you to really visualize the tests and helps you generate the syntax you need to create them.

So far, I have only created a test that accesses the create-list page, then tests that when a user clicks the "create shopping list" button, the user is redirected to the list page.