Add radio buttons to add item form
We've added radio buttons to the form so that the user can set how soon they plan on purchasing that item again. Using React hooks, we've updated the state to include a duration so that when a radio button is selected and the form is submitted, that duration value will be sent to the database.